New optional postflight and per-field interfaces. Fixed a missing-tag bug.
PostFlighter is an interface which, if implemented for a scraper, will
perform post-flight operations on the scraped object. It can be passed
arbitrary context from the entrypoint functions. This could be used to
use scraped data to construct conditional or accumulative fields.

FieldSelecter is a per-field method that will be called with the fieldName
and the context provided to the entrypoint functions, and can return
whether or not to fill the field. For example, a datatype that has fields
that should be filled from separate pages could pass the current URL as
context, and have a method that skips fields not relevant to the current
URL.

There was also a bug where fields that were missing `csss:""` tags were
leading to scrape failure; I had falsely assumed that only fields with
such tags were getting returned by reflections.